refactor: Move CSV parsing helper methods to a new class for better organization
